app/inference.py
"""
Inference logic for CIFAR-10 classification
"""

import torch
from PIL import Image
import numpy as np


class CIFAR10Predictor:
    """
    Predictor class for CIFAR-10 classification
    """
    
    def __init__(self, model_path, device='cpu'):
        """
        Initialize predictor
        
        Args:
            model_path: Path to saved model
            device: Device to run inference on
        """
        self.device = device
        self.model = None
        self.load_model(model_path)
    
    def load_model(self, model_path):
        """
        Load trained model
        
        Args:
            model_path: Path to saved model
        """
        pass
    
    def predict(self, image):
        """
        Predict class for input image
        
        Args:
            image: Input image (PIL Image or numpy array)
            
        Returns:
            Predicted class and confidence scores
        """
        pass
    
    def predict_batch(self, images):
        """
        Predict classes for batch of images
        
        Args:
            images: List of images
            
        Returns:
            List of predictions
        """
        pass
