requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image Processing
Pillow>=10.0.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
